requirements.txt
pyyaml>=6.0.3
fire>=0.7.1
tqdm>=4.65.0
google-genai>=1.69.0
google-generativeai>=0.8.6